Move attachment command into ui.commands
